+++ docs/source/advanced/crude_datasets.md
@@ -34,8 +34,8 @@ subflavors:
 ```
 
 As shown above, [subflavors](sect-subflavors) can be assigned directly in the dataset yaml, however we recommend
-to define them in the [Metadataset](../basic/metadataset) instead.
-Subflavors can be used to differentiate the data origin when you mix different crude datasets or crude with non-crude data using a [Metadataset](../basic/metadataset).
+to define them in the [Recipe](../basic/recipe) instead.
+Subflavors can be used to differentiate the data origin when you mix different crude datasets or crude with non-crude data using a [Recipe](../basic/recipe).
 You can then use the subflavors to determine how each sample shall be processed, as explained below.
 
 Let's see how we set up our [Task Encoder](../basic/task_encoder) to cook the crude samples and turn them into real samples like {py:class}`TextSample <megatron.energon.TextSample>`.
@@ -121,10 +121,10 @@ An auxiliary data source can be either
 * Another energon-prepared WebDataset
 * A folder on the local or a remote file system
 
-You can specify it in your [metadataset](../basic/metadataset) yaml as follows (look at the `aux:` section)
+You can specify it in your [recipe](../basic/recipe) yaml as follows (look at the `aux:` section)
 ```yaml
 __module__: megatron.energon
-__class__: MetadatasetV2
+__class__: Recipe
 splits:
   train:
     path: ./my_pimary_ds
@@ -172,7 +172,7 @@ def cook_text(sample: dict, foo_bar_source: FileStore) -> TextSample:
 # ...
 ```
 
-You can use multiple sources. You'll have to specify a cooker argument for each source that was defined in the metadataset.
+You can use multiple sources. You'll have to specify a cooker argument for each source that was defined in the recipe.
 
 For easier debugging, you should always keep track of all the sources you used. The `get` method takes care of this if you pass it the sample like this:
 
@@ -183,7 +183,7 @@ additional_text = foo_bar_source.get(sample['add_txt_fname'], sample)
 This will update the sample-internal `__sources__` list with the aux dataset you used.
 
 For byte-range auxiliary data, use a key of the form `path/to/file#bytes=offset:size`.
-The path is resolved below the `byterange://` root from the metadataset.
+The path is resolved below the `byterange://` root from the recipe.
 
 ```python
 def cook_text(sample: dict, byte_range_source: FileStore) -> TextSample:
