+++ docs/source/advanced/epochized_blending.md
@@ -6,11 +6,11 @@ SPDX-License-Identifier: BSD-3-Clause -->
 As an alternative to blending with a weight for each dataset, blending can be made accurate and
 iterating the dataset can follow epochs (i.e. interrupt iteration after an epoch) with this concept.
 
-Here is an example `metadataset.yaml` config file that changes to epochized blending:
+Here is an example `recipe.yaml` config file that changes to epochized blending:
 
 ```yaml
 __module__: megatron.energon
-__class__: MetadatasetV2
+__class__: Recipe
 splits:
   train:
     # Blend the following datasets, repeating coco 5 times, coyo-train 2 times and coyo-val 1 times
@@ -32,7 +32,7 @@ from megatron.energon import get_train_dataset, get_loader, WorkerConfig
 
 
 loader = get_loader(get_train_dataset(
-    'metadataset.yaml',
+    'recipe.yaml',
     batch_size=2,
     shuffle_buffer_size=100,
     max_samples_per_sequence=100,
@@ -53,4 +53,4 @@ for batch in loader:
 ```
 
 If used as dataset for `get_val_dataset`, the `repetitions` are ignored.
-The metadataset would also work without setting `repeat=False`, but then the shuffle buffer will shuffle samples across bounderies of epochs.
+The recipe would also work without setting `repeat=False`, but then the shuffle buffer will shuffle samples across bounderies of epochs.
