+++ docs/source/advanced/remote_dataset.md
@@ -60,7 +60,7 @@ msc://coolstore/mainbucket/datasets/somedata
 You can use this URL instead of paths to datasets in
 
 * Functions like `get_train_dataset`, `get_val_dataset`
-* Inside [metadataset](../basic/metadataset) specifications
+* Inside [recipe](../basic/recipe) specifications
 * As arguments to `energon prepare` or `energon lint`. Note that those may be slow for remote locations.
 * Or as a path to [`energon mount`](energon-mount) to locally inspect your remote dataset 😎
