+++ docs/source/basic/data_prep.md
@@ -71,7 +71,7 @@ These are the typical steps to get your data ready:
 3. Run our preparation tool `energon prepare` **on both datasets** (yes also on JSONL or Parquet directories) to convert to an energon-compatible format
     * Configure WebDataset-based crude datasets as `CrudeWebdataset`; JSONL and Parquet datasets are loaded as crude data by default.
     * For the auxiliary datasets, we recommend to enable the [media metadata feature](media-metadata) to store additional information about the media (like image size, resolution, video duration etc.)
-4. Create a [metadataset](../basic/metadataset) that specifies what auxiliary data to load for each primary dataset
+4. Create a [recipe](../basic/recipe) that specifies what auxiliary data to load for each primary dataset
     * For more details read about [crude data](crude-data)
 
 (create-jsonl-dataset)=
@@ -106,7 +106,7 @@ each file containing the JSON from one of the lines above.
 
 None of the JSON fields is mandatory. The data is considered to be crude data and will be interpreted by your custom [cooker](crude-data).
 If you want to include media, you should include file names of the media files in the JSON.
-A metadataset with [auxiliary data](aux-data) can then be used to load the media on the fly.
+A recipe with [auxiliary data](aux-data) can then be used to load the media on the fly.
 
 Here's an example of how a polylithic JSONL dataset with images might look like:
 
@@ -119,14 +119,14 @@ Steps needed:
 
 1. Create the JSONL file according to your needs
 2. Run `energon prepare /path/to/my_dataset.jsonl` to create an index next to it
-3. Optionally create a [metadataset](../basic/metadataset) that specifies what auxiliary data to load for each primary dataset
+3. Optionally create a [recipe](../basic/recipe) that specifies what auxiliary data to load for each primary dataset
     * For more details read about [crude data](crude-data)
 
-The metadataset would then refer to the JSONL dataset while specifying the auxiliary data source:
+The recipe would then refer to the JSONL dataset while specifying the auxiliary data source:
 
 ```yaml
 __module__: megatron.energon
-__class__: MetadatasetV2
+__class__: Recipe
 splits:
   train:
     path: /path/to/my_dataset.jsonl
@@ -369,7 +369,7 @@ These are the possible integrated types you can currently choose from:
   * Attributes:
     * {py:attr}`__key__: str <megatron.energon.Sample.__key__>`: Unique identifier of the sample within the dataset. Useful for backtracking the source of a single sample.
     * {py:attr}`__key__: str <megatron.energon.Sample.__restore_key__>`: Structured key of the sample, which can be used to regenerate the sample without storing the whole sample.
-    * {py:attr}`__subflavors__: dict[str, Any] | None <megatron.energon.Sample.__subflavors__>`: Represents the subflavors (i.e. custom dict data) set for the source dataset (typically in the metadataset).
+    * {py:attr}`__subflavors__: dict[str, Any] | None <megatron.energon.Sample.__subflavors__>`: Represents the subflavors (i.e. custom dict data) set for the source dataset (typically in the recipe).
   * {py:class}`CaptioningSample <megatron.energon.CaptioningSample>`: Represents a sample for captioning
     * Attributes:
       * {py:attr}`image: torch.Tensor <megatron.energon.CaptioningSample.image>`: The input image tensor
