+++ docs/source/internals/contrib_guidelines.md
@@ -12,9 +12,9 @@ If you want to contribute to this repository please adhere to the following guid
 - User-exposed classes and methods should be documented in Google-style docstrings that are parsed by sphinx
   and end up in this documentation
 - Breaking changes should be marked in the message of pull requests:
-  - `CHECKPOINT BREAKING CHANGE`: When the save/restore structure changed incompatibly (check test `test_metadataset:TestDataset.test_save_restore_state_train`)
+  - `CHECKPOINT BREAKING CHANGE`: When the save/restore structure changed incompatibly (check test `test_recipe:TestDataset.test_save_restore_state_train`)
   - `ITERATION ORDER BREAKING CHANGE`: When the order of iterating samples changed, i.e. experiments would not be exactly reproducible (check tests `test_dataset:TestDataset.test_current_batch_index_generator`, `test_dataset:TestDataset.test_current_batch_index`, maybe more)
   - `API BREAKING CHANGE`: When the external programming api changed incompatibly
   - `DATASET CONFIG BREAKING CHANGE`: When the dataset config (`.nv-meta` folder) changed incompatibly
-  - `METADATASET CONFIG BREAKING CHANGE`: When the metadataset config changed
+  - `RECIPE CONFIG BREAKING CHANGE`: When the recipe config changed
 - In a release, all breaking changes except checkpoint lead to a new major version.
