+++ src/megatron/energon/recipe/loader_interface.py
@@ -45,14 +45,14 @@ class LoadedDatasetList:
 
 @dataclass
 class TraversedDatasetReference:
-    """Flattened leaf dataset reference produced by metadataset traversal.
+    """Flattened leaf dataset reference produced by recipe traversal.
 
     Attributes:
         path: Resolved path to the referenced leaf dataset.
         split_part: Effective split part to use when loading the leaf dataset.
         aux: Resolved auxiliary dataset or filesystem references keyed by auxiliary name.
-        subflavors: Effective subflavors implied by the traversed metadataset hierarchy.
-        shuffle_over_epochs_multiplier: Effective shuffle over epochs multiplier from metadataset references.
+        subflavors: Effective subflavors implied by the traversed recipe hierarchy.
+        shuffle_over_epochs_multiplier: Effective shuffle over epochs multiplier from recipe references.
     """
 
     path: EPath
@@ -66,33 +66,33 @@ class DatasetLoaderInterface(ABC):
     """General interface for a dataset loader."""
 
     @abstractmethod
-    def post_initialize(self, mds_path: Optional[EPath] = None):
+    def post_initialize(self, recipe_path: Optional[EPath] = None):
         """Called to finally initialize the dataset."""
         ...
 
     def traverse(
         self,
-        mds_path: Optional[EPath] = None,
+        recipe_path: Optional[EPath] = None,
         *,
         split_part: Union[Literal["train", "val", "test"], str],
         _shuffle_over_epochs_multiplier: Optional[int] = 1,
         _subflavors: Optional[Dict[str, Any]] = None,
     ) -> List[TraversedDatasetReference]:
-        """Traverse a metadataset subtree and collect flattened leaf dataset references.
+        """Traverse a recipe subtree and collect flattened leaf dataset references.
 
         This method is the traversal-side counterpart to `get_datasets()`. Instead of
         instantiating dataset loaders for leaf datasets, it walks the hierarchy, resolves nested
-        metadataset references, and returns the final leaf dataset references for a single split.
+        recipe references, and returns the final leaf dataset references for a single split.
 
         Args:
-            mds_path: Parent metadataset path used internally to resolve relative dataset and
+            recipe_path: Parent recipe path used internally to resolve relative dataset and
                 auxiliary paths. Must be set for nested references and inner traversal nodes;
-                use None only for top-level metadatasets.
+                use None only for top-level recipes.
             split_part: Split to traverse, such as `\"train\"`, `\"val\"`, or `\"test\"`. Nested
                 references may override this with their own configured split.
             _shuffle_over_epochs_multiplier: Inherited shuffle multiplier (merged per node like
                 ``get_datasets``); default ``1``.
-            _subflavors: Effective subflavors implied by the traversed metadataset hierarchy.
+            _subflavors: Effective subflavors implied by the traversed recipe hierarchy.
 
         Returns:
             A flattened list of `TraversedDatasetReference` values for all leaf datasets reached
